requirements.txt
asgiref==3.2.3
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.10
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-api==0.1.5
django-rest-auth==0.9.5
django-toolbelt==0.0.1
djangorestframework==3.11.0
drf-yasg==1.17.1
environ==1.0
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
traitlets==4.3.3
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
